refactor: split admin tools router by domain
- replace backend/routers/admin_tools.py with admin_nfl, admin_live_scoring,
  admin_drafts, and admin_config routers
- register split routers in backend/main.py and remove the combined router
- move Site Admin NFL calls to /admin/nfl/* and update Vite proxy entries
- update backend tests and frontend smoke coverage for the new paths
- refresh docs/runbooks to publish the new admin endpoint layout
