+++ docs/API_PAGE_MATRIX.md
@@ -32,7 +32,7 @@ Base URL (frontend client): `http://127.0.0.1:8000`
 | Trades               | `POST /trades/propose`, `GET /trades/pending`, `POST /trades/{trade_id}/approve`, `POST /trades/{trade_id}/reject`                                                                                                | My Team, Commissioner                                 |
 | Feedback             | `POST /feedback/bug`                                                                                                                                                                                                | Bug Report                                            |
 | Advisor (AI wrapper) | `GET /advisor/status`, `POST /advisor/ask`                                                                                                                                                                          | Global advisor widget                                 |
-| Admin tools          | `POST /admin/tools/sync-nfl`, `POST /admin/tools/import-nfl-schedule`, `POST /admin/tools/uat-draft-reset`, `POST /admin/tools/uat-team-reset`, `POST /admin/tools/reload-config`, `GET/POST/PUT/DELETE /admin/tools/commissioners`, `POST /admin/create-test-league`, `POST /admin/reset-draft`, `POST /admin/finalize-draft`, `POST /admin/reset-league` | Site Admin                                            |
+| Admin tools          | `POST /admin/nfl/sync`, `POST /admin/nfl/schedule/import`, `POST /admin/live-scoring/ingest`, `GET /admin/live-scoring/health`, `POST /admin/live-scoring/watchdog`, `GET /admin/live-scoring/watchdog/alerts`, `POST /admin/drafts/refresh-values`, `POST /admin/config/reload`, `POST /admin/tools/uat-draft-reset`, `POST /admin/tools/uat-team-reset`, `GET/POST/PUT/DELETE /admin/tools/commissioners`, `POST /admin/create-test-league`, `POST /admin/reset-draft`, `POST /admin/finalize-draft`, `POST /admin/reset-league` | Site Admin                                            |
 
 ## External APIs (third-party)
 
@@ -43,7 +43,7 @@ Base URL (frontend client): `http://127.0.0.1:8000`
 | ESPN public NFL endpoints | `https://site.api.espn.com/apis/site/v2/sports/football/nfl/*`          | Player/stats ingestion scripts        | `backend/scripts/import_espn_players.py`, `backend/scripts/archive_weekly_stats.py` |
 | Yahoo Fantasy API         | `https://football.fantasysports.yahoo.com/f1/draftanalysis?type=salcap` | Draft value ingestion                 | `backend/scripts/import_yahoo_players.py`                                           |
 | Draftsharks ADP           | `https://www.draftsharks.com/adp/superflex/ppr/sleeper/12`              | Draft value ingestion                 | `backend/scripts/import_draftsharks_players.py`                                     |
-| FantasyNerds Auction API  | `https://api.fantasynerds.com/v1/nfl/auction/`                          | Auction values with min/max context   | `POST /admin/tools/refresh-draft-values` via `etl/extract/extract_fantasynerds.py` |
+| FantasyNerds Auction API  | `https://api.fantasynerds.com/v1/nfl/auction/`                          | Auction values with min/max context   | `POST /admin/drafts/refresh-values` via `etl/extract/extract_fantasynerds.py` |
 | RubeSheets (optional)     | `https://rubesheets.com/Footballv2.aspx`                                | League-custom auction modeling option | Not enabled; documented fallback option only (WebForms scrape risk)                 |
 
 ## Draft Value API & Page Mapping
@@ -69,7 +69,7 @@ Draft value data is sourced from ESPN, Yahoo, and Draftsharks APIs, normalized, 
 Operational notes:
 
 - Weekly source prechecks run via `.github/workflows/source-prechecks.yml`.
-- `POST /admin/tools/refresh-draft-values` now supports fail-closed prechecks for Yahoo and FantasyNerds.
+- `POST /admin/drafts/refresh-values` now supports fail-closed prechecks for Yahoo and FantasyNerds.
 - RubeSheets is intentionally not integrated into production flow at this time due to brittle form-post parsing and maintenance risk.
 
 ---
@@ -95,7 +95,7 @@ Notes:
 | `/commissioner` Commissioner Dashboard + modals | `GET /leagues/{leagueId}/settings` (includes waiver rules and budget metadata), `GET /leagues/{leagueId}/waiver-budgets`, `GET /leagues/owners?league_id={leagueId}`, `PUT /leagues/{leagueId}/settings`, `POST /leagues/owners`, `GET /trades/pending`, `POST /trades/{tradeId}/{action}` (frontend reference), `POST /leagues/{leagueId}/draft-year`, `POST /leagues/{leagueId}/budgets` |
 | `/commissioner/manage-divisions`               | `GET /leagues/owners?league_id={leagueId}`, `GET /leagues/{leagueId}/divisions/config`, `PUT /leagues/{leagueId}/divisions/config`, `POST /leagues/{leagueId}/divisions/assignment-preview`, `POST /leagues/{leagueId}/divisions/finalize`, `POST /leagues/{leagueId}/divisions/undo-last`, `POST /leagues/{leagueId}/divisions/report-name` |
 | `/commissioner/history-owner-mapping`          | `GET /leagues/owners?league_id={leagueId}`, `GET /leagues/{leagueId}/history/team-owner-map`, `PUT /leagues/{leagueId}/history/team-owner-map`, `DELETE /leagues/{leagueId}/history/team-owner-map/{rowId}`, `GET /leagues/{leagueId}/history/unmapped-series-keys`, `GET /leagues/{leagueId}/history/owner-gap-report` |
-| `/admin` Site Admin                             | `POST /admin/tools/sync-nfl`, `POST /admin/tools/import-nfl-schedule`, `POST /admin/tools/uat-draft-reset`, `POST /admin/tools/uat-team-reset`, `POST /admin/create-test-league`, `POST /admin/reset-draft`                                                                                                                                                                              |
+| `/admin` Site Admin                             | `POST /admin/nfl/sync`, `POST /admin/nfl/schedule/import`, `POST /admin/tools/uat-draft-reset`, `POST /admin/tools/uat-team-reset`, `POST /admin/create-test-league`, `POST /admin/reset-draft`                                                                                                                                                                                           |
 | `/admin/manage-commissioners`                   | `GET /admin/tools/commissioners`, `POST /admin/tools/commissioners`, `PUT /admin/tools/commissioners/{commissionerId}`, `DELETE /admin/tools/commissioners/{commissionerId}`                                                                                                                                                                                                             |
 | `/bug-report` Bug Report                        | `PUT /auth/email`, `POST /feedback/bug`                                                                                                                                                                                                                                                                                                                                                    |
 | `/analytics` Analytics Dashboard                | Uses `/analytics/league/{id}/leaderboard`, `/analytics/league/{id}/weekly-stats` and `/analytics/league/{id}/rivalry` endpoints to power charts and tables (efficiency, weekly trends, rivalry graph)                                                                                                                                                                                                                                           |
